feat: reviewModal refactored
